Repaired integrations tests; replaced StreamWriter output with List<GCode>
a) The previous change (moving Z words always to the last
position, because uCNC+UGS do not work with inner comments)
left some integrations tests red - these are now green.

b) In preparation for a peephole optimizer, the string
output of the Emit phase is now replaced with a list of
GCode objects.