Added optimizer pass and implemented sweep peephole optimizer
Multiple back-to-back sweeps in different directions are
almost always useless: A single sweep to the endpoint is
all that is needed. To save some time, especially in
scenarios with multiple subpaths, I have now added an
optimizer pass, which eliminates multiple sweeps via
a simple peephole optimization.

Unfortunately, this required a rewrite of the output
generation - it was via strings before, now it is to
special GCode objects. Also, I had to move the
statistics aggregation after the optimizer pass, which
changed a lot of method signatures.