Better defaults for S values; bugfix
a) Up to now, the sweep height was
determined quite clumsily (explicit
definitions are necessary).
From now on, the default S for the
main path must be provided as /s
parameter (unless /c is specified);
this S value must clear all setup
hurdles in the working area.
For the subpaths, the default is
T + O, i.e. "a little but above T";
this assumes that no fastening
objects are present in the work
area of a subpath.
This should in almost all cases
allow to drop the S values at
paths; and hence reduce the risk
of erroneous S specifications.

b) Fixed an error with Z value
which prevented some optimizations.