Added diameter limit for helical circles
Major changes
- Helical circles draw the outer diameter
of the milled circle, not the path of the
milling bit. This is a risky difference
to all other paths; therefore, I have now
limited the use of helical circles to
"smaller circles" ("holes"). The default
maximum is 4 x O; this can be changed
with the A parameter on the path.

Small changes
- NONE

Refactorings:
- NONE