New parameter for probing speed
Major changes
- Probing ran with F speed unless
Z was specified at every path with
Z probes. Of course, I forgot that,
and probed too fast, i.e. the bit
dug into the aluminum.
Now, there is a /z parameter for the
probing speed. It is non-required,
and its default is the /f value - but
one can add it to a batch file, and it
remains there forever :-). My default
is /z50 (and /f150), which works out
nicely.

Small changes
- No changes

Refactorings
- No changes