PathGCodeAdjustZ outputs some lines
Major changes
- In the gcode file, there is some
useful information about milling
lengths and times. Instead of
opening the file(s) with an editor
and searching for those lines, it
is now possible to pass a regex
PathGCodeAdjustZ as /x, and
then matching lines are written
to stdout. I use the pattern
mm.*min to see the statistics
lines.

Small changes
- Improved length checking for
subpath transformation.

Refactorings
- No changes