Major rewrite for parameterization of models
Major changes
- Often, milling jobs are very similar. For
example, in one job, one might create items
of type A and B, whereas in another job,
compatible items of type C and D should be
milled. It is useful to have a common file
for milling two such items, and then just
parameterize this drawing with A,B or C,D.
This concept requires a much clearer
distinction between RawPathModels (where
the variable values are not yet relevant)
and PathModels, where variables are
expanded and hence values and even subpaths
might be different depending on the passed
variables.
This commit contains almost all the
necessary archtectural changes. It is not
yet tested (and hence, the two basic
integration tests are red in this commit);
also, the envisioned constraints for
variables are not yet implemented (they
are needed for two purposes: (a) a
documentation which values can be passed
into a parameterized model; and (b) for
checking with -c, variable values have to
be generated by PathDxf2GCode.
The goal of this commit is solely to lay
the foundations for this change, without
any changes to the current behavior (with
the exception of the changes indicated
below).

Small changes, bugfixes
- Matching of filenames to a pathnames for
searching files to be read somewhere became
haphazard. It has now been reverted more or
less to the initial design, with an
additional /n option for pathname regexes.

Refactorings
- Essentially, the changes described under
"Major changes" are a huge refactoring.