Microscopic change aftet asking Copilot
Major changes
- None.

Small changes, bugfixes
- Just for fun, I asked Copilot "Tell
me all errors you find". The result is
as I expected:
a) It told me multiple times "this is
correct" and "that is robust". I knew
that anyway ...
b) It told me a change to a field; I've
changed it just to be able to commit
this.
c) And it told me multiple errors in
IntegrationTests.cs ... which are simply
not there; e.g., it said "The file
IntegrationTests.cs has syntax errors and
will not compile" - but it compiles :-).
Summing up: Copilot is useless and dumb.

Refactorings
- None.

Open issues
- None.